feat: Added AsideBar
