+++ src/contexts/СhatDataContext/index.tsx
@@ -1,5 +1,5 @@
 import { getFormattedTime } from "@utils/constants";
-import { useState, useEffect, useContext, createContext } from "react";
+import { useState, useEffect, createContext } from "react";
 import {
     ClientMessagesTypes,
     type ITextMessage,
@@ -29,7 +29,7 @@ interface IChatDataContext {
     messageHandlersConfig: IMessageHandlerData[];
 }
 
-const ChatDataContext = createContext<IChatDataContext | null>(null);
+export const ChatDataContext = createContext<IChatDataContext | null>(null);
 
 export const ChatDataProvider: React.FC<{
     children: React.ReactNode;
@@ -134,14 +134,4 @@ export const ChatDataProvider: React.FC<{
             {children}
         </ChatDataContext.Provider>
     );
-};
-
-export const useChatDataContext = () => {
-    const context = useContext(ChatDataContext);
-    if (!context) {
-        throw new Error(
-            "useChatDataContext must be used within ChatDataProvider"
-        );
-    }
-    return context;
 };
